requirements.txt
Django==1.5
Pillow==1.7.8
South==0.7.6
argparse==1.2.1
beautifulsoup4==4.1.3
django-appconf==0.6
django-braces==1.0.0
django-class-based-auth-views==0.1a2
django-compressor==1.2
django-crispy-forms==1.2.3
django-debug-toolbar==0.9.4
django-extensions==1.0.3
django-imagekit==2.0.3
django-social-auth==0.7.20
httplib2==0.7.7
ipython==0.13.1
oauth2==1.5.211
psycopg2==2.4.6
python-openid==2.2.5
python-social-auth==0.1.8
requests==1.1.0
six==1.2.0
wsgiref==0.1.2
